Pull the comparators out of MessageListFragment
